test(processing): #2689 consolidate monitor tests + add BoundedContext trait
